+++ FinBud-Frontend/src/_components/Calculator-Component/CalculatorComponent.tsx
