+++ FinBud-Frontend/src/_components/Calculator-Component/Reference.tsx
