+++ FinBud-Frontend/src/_components/Responsive-Image-Component/ResponsiveImage.tsx
