--- FinBud-Frontend/src/types.d.ts
@@ -1,11 +1,0 @@
-export interface Answer {
-  text: string;
-  nextQuestionId?: number;
-  result?: string;
-}
-
-export interface Question {
-  id: number;
-  question: string;
-  answers: Answer[];
-}
